\name{SGPmatrices-package}
\alias{SGPmatrices-package}
\docType{package}
\title{Baseline coefficient matrices for SGP analyses}
\description{
	The package SGPmatrices contains baseline coefficient matrices for multiple states/assessment for SGP analyses.
}

\details{
\tabular{ll}{
Package: \tab SGPmatrices\cr
Type: \tab Package\cr
Version: \tab 0.1-5.0\cr
Date: \tab 2026-9-4\cr
License: \tab GPL-3\cr
LazyLoad: \tab Yes\cr
LazyData: \tab Yes\cr
}
}

\keyword{package}
