\name{SGPmatrices}
\alias{SGPmatrices}
\docType{data}
\title{Baseline matrices for baseline SGP analyses}
\description{
An environment containing collections of baseline matrices used historically during SGP baseline analyses. Matrices might be updated periodically.
Matrices are used if baseline SGP analyses are performed. The environment currently contains matrices associated with the following states/organizations/years:

\describe{
    \item{\code{WIDA_IN_Baseline_Matrices}}{Contains 2021 baseline matrices for WIDA_IN}
}
}

\source{Large Scale Summative Assessment Data}
\keyword{datasets}
