\name{WIDA_IN_Baseline_Matrices}
\alias{WIDA_IN_Baseline_Matrices}
\docType{data}
\title{Baseline matrices for WIDA Indiana baseline SGP analyses}
\description{
An environment containing collections of baseline matrices used historically during SGP baseline analyses. Matrices might be updated periodically.
Matrices are used if baseline SGP analyses are performed for WIDA_IN (WIDA-ACCESS Indiana). The environment currently contains matrices associated with the following years:

\describe{
    \item{\code{WIDA_IN_Baseline_Matrices_2021}}{2021 WIDA_IN Baseline SGP Analyses}
}
}

\source{WIDA-ACCESS Indiana Assessment Data}
\keyword{datasets}
