\name{SGPmatrices}
\alias{SGPmatrices}
\docType{data}
\title{Baseline matrices for baseline SGP analyses}
\description{
An environment containing collections of baseline matrices used historically during SGP baseline analyses. Matrices might be updated periodically.
Matrices are used if baseline SGP analyses are performed. The environment currently contains matrices associated with the following states/organizations/years:

\describe{
    \item{\code{AZ_Baseline_Matrices}}{Contains 2021 baseline matrices for Arizona AzM2}
    \item{\code{CO_Baseline_Matrices}}{Contains 2021 baseline matrices for Colorado CMAS}
    \item{\code{HI_Baseline_Matrices}}{Contains 2021 baseline matrices for Hawaii SBA}
    \item{\code{ID_Baseline_Matrices}}{Contains 2021 baseline matrices for Idaho ISAT}
    \item{\code{IN_Baseline_Matrices}}{Contains 2021 baseline matrices for Indiana ILEARN/ISTEP+}
    \item{\code{MA_Baseline_Matrices}}{Contains 2021 baseline matrices for Massachusetts MCAS 2.0/PARCC}
    \item{\code{MI_Baseline_Matrices}}{Contains 2021 baseline matrices for Michigan MSTEP/PSAT/SAT}
    \item{\code{NH_Baseline_Matrices}}{Contains 2021 baseline matrices for New Hampshire SAS}
    \item{\code{RI_Baseline_Matrices}}{Contains 2021 baseline matrices for Massachusetts RICAS/PARCC}
    \item{\code{UT_Baseline_Matrices}}{Contains 2021 baseline matrices for Utah RISE and Utah Aspire Plus}
    \item{\code{WA_Baseline_Matrices}}{Contains 2021 baseline matrices for Washington Smarter-Balanced Assessment}
    \item{\code{WIDA_Baseline_Matrices}}{Contains 2021 baseline matrices for multiple WIDA members}
    \item{\code{WIDA_CO_Baseline_Matrices}}{Contains 2021 baseline matrices for Colorado WIDA-ACCESS}
    \item{\code{WIDA_HI_Baseline_Matrices}}{Contains 2021 baseline matrices for Hawaii WIDA-ACCESS}
    \item{\code{WIDA_ID_Baseline_Matrices}}{Contains 2021 baseline matrices for Idaho WIDA-ACCESS}
    \item{\code{WIDA_IN_Baseline_Matrices}}{Contains 2021 baseline matrices for Indiana WIDA-ACCESS}
    \item{\code{WIDA_MA_Baseline_Matrices}}{Contains 2021 baseline matrices for Massachusetts WIDA-ACCESS}
    \item{\code{WIDA_NH_Baseline_Matrices}}{Contains 2021 baseline matrices for New Hampshire WIDA-ACCESS}
    \item{\code{WIDA_NM_Baseline_Matrices}}{Contains 2021 baseline matrices for New Hampshire WIDA-ACCESS}
    \item{\code{WIDA_RI_Baseline_Matrices}}{Contains 2021 baseline matrices for Rhode Island WIDA-ACCESS}
}
}

\source{Large Scale Summative Assessment Data}
\keyword{datasets}
